Put the expand_tilde test at the end of the file
